Move `:core:data` fake repositories to `:core:data-test`
Now that these fake implementations are no longer used in the production source code (only in tests), we can safely extract them out of this module.
Hopefully, we'll later be able to use Kotlin testFixtures for that.

References:
- #709
- #875
