+++ core/data-test/src/main/kotlin/com/google/samples/apps/nowinandroid/core/data/test/TestDataModule.kt
@@ -22,11 +22,11 @@ import com.google.samples.apps.nowinandroid.core.data.repository.RecentSearchRep
 import com.google.samples.apps.nowinandroid.core.data.repository.SearchContentsRepository
 import com.google.samples.apps.nowinandroid.core.data.repository.TopicsRepository
 import com.google.samples.apps.nowinandroid.core.data.repository.UserDataRepository
-import com.google.samples.apps.nowinandroid.core.data.repository.fake.FakeNewsRepository
-import com.google.samples.apps.nowinandroid.core.data.repository.fake.FakeRecentSearchRepository
-import com.google.samples.apps.nowinandroid.core.data.repository.fake.FakeSearchContentsRepository
-import com.google.samples.apps.nowinandroid.core.data.repository.fake.FakeTopicsRepository
-import com.google.samples.apps.nowinandroid.core.data.repository.fake.FakeUserDataRepository
+import com.google.samples.apps.nowinandroid.core.data.test.repository.FakeNewsRepository
+import com.google.samples.apps.nowinandroid.core.data.test.repository.FakeRecentSearchRepository
+import com.google.samples.apps.nowinandroid.core.data.test.repository.FakeSearchContentsRepository
+import com.google.samples.apps.nowinandroid.core.data.test.repository.FakeTopicsRepository
+import com.google.samples.apps.nowinandroid.core.data.test.repository.FakeUserDataRepository
 import com.google.samples.apps.nowinandroid.core.data.util.NetworkMonitor
 import dagger.Binds
 import dagger.Module
@@ -38,7 +38,7 @@ import dagger.hilt.testing.TestInstallIn
     components = [SingletonComponent::class],
     replaces = [DataModule::class],
 )
-interface TestDataModule {
+internal interface TestDataModule {
     @Binds
     fun bindsTopicRepository(
         fakeTopicsRepository: FakeTopicsRepository,
